Segregate api v0.0
