+++ server/api/__test__/api.v0.0.spec.js
@@ -1,0 +1,9 @@
+// https://stackoverflow.com/questions/24153261/joining-tests-from-multiple-files-with-mocha-js
+
+describe('API v0.0', () => {
+  require('../v0.0/test/api.auth.login.spec.js');
+  require('../v0.0/test/api.auth.signup.spec.js');
+  require('../v0.0/test/api.item.spec.js');
+  require('../v0.0/test/api.list.spec.js');
+  require('../v0.0/test/api.user.spec.js');
+});
