+++ server/api/v0.0/test/api.v1.0.spec.js
@@ -1,6 +1,8 @@
 // https://stackoverflow.com/questions/24153261/joining-tests-from-multiple-files-with-mocha-js
 
-describe('API v1.0', () => {
+// const loginTest = require('../v1.0/test/api.auth.login.spec.js');
+
+describe('api v1.0', () => {
   require('../v1.0/test/api.auth.login.spec.js');
   require('../v1.0/test/api.auth.signup.spec.js');
   require('../v1.0/test/api.item.spec.js');
